+++ ccramic/parsers/pixel_level_parsers.py
@@ -4,7 +4,6 @@ from pathlib import Path
 from tifffile import TiffFile
 import os
 from ccramic.utils.pixel_level_utils import (
-    convert_to_below_255,
     split_string_at_pattern,
     set_array_storage_type_from_config)
 from readimc import MCDFile, TXTFile
@@ -194,7 +193,7 @@ def populate_upload_dict(uploaded_files, array_store_type="float"):
                             for image in acq:
                                 image_label = txt_channel_labels[image_index - 1]
                                 identifier = txt_channel_names[image_index - 1]
-                                upload_dict[roi][identifier] = convert_to_below_255(image).astype(
+                                upload_dict[roi][identifier] = image.astype(
                                     set_array_storage_type_from_config(array_store_type))
                                 if image_index == 1:
                                     dataset_information["ROI"].append(str(roi))