+++ ccramic/utils/pixel_level_utils.py
@@ -127,8 +127,8 @@ def get_area_statistics_from_closed_path(array, svgpath):
     return np.average(array[masked_array]), np.max(array[masked_array]), np.min(array[masked_array])
 
 
-def convert_to_below_255(array):
-    return array if np.max(array) < 65000 else (array // 256).astype(np.uint8)
+# def convert_to_below_255(array):
+#     return array if np.max(array) < 65000 else (array // 256).astype(np.uint8)
 
 
 def resize_for_canvas(image, basewidth=400, return_array=True):